REFACTOR: use subpackage for Distance
